update lib/pq from 1.1.1 to 1.10.2 (latest)
I'm seeing "broken pipe" errors randomly on one of our servers
The issue seems to be the db driver conns become stale when tcp connections were disconnected.

It happens more often when the DB is behind a proxy.

This has been reported and fixed in the lib/pq in v1.9+

https://github.com/lib/pq/pull/871